backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
openai
pytest
pytest-mock
beautifulsoup4
python-multipart